Switch CI to retrieve baremetal servers from a resource pool
Move to a resource pooler that reuses servers instead of allocating
them fresh each time. This method should be faster and less likely
to hit capacity errors. Also can be extended to other cloud
providers.

If the new method fails then fallback to the ansible method.

+++ ci-operator/step-registry/baremetalds/packet/teardown/baremetalds-packet-teardown-commands.sh
@@ -6,6 +6,14 @@ set -o pipefail
 
 echo "************ baremetalds packet teardown command ************"
 
+set -x
+CIRFILE=$SHARED_DIR/cir
+if [ -e $CIRFILE ] ; then
+    curl -kfX DELETE -H "Host: ofcir.apps.ostest.test.metalkube.org" "$(cat ${CLUSTER_PROFILE_DIR}/ofcir_url)/$(jq -r .name < $CIRFILE)" || true
+    exit 0
+fi
+set +x
+
 # Run Ansible playbook
 cd
 cat > packet-teardown.yaml <<-EOF
